consolidate and add psuedocode
